Use decimal for currencies, add gitignore, replace urllib by requests

+++ api/models.py
@@ -16,15 +16,15 @@ class Currency(models.Model):
 class Rate(models.Model):
     base = models.ForeignKey(Currency, on_delete=models.CASCADE, related_name='rates')
     code = models.CharField(max_length=3, choices=CURRENCY_CHOICES, blank=False)
-    rate = models.FloatField(blank=False)
+    rate = models.DecimalField(blank=False, decimal_places=12, max_digits=16)
 
 
 class ConverterResponse(models.Model):
 
     base = models.CharField(max_length=3, choices=CURRENCY_CHOICES, blank=False)
     target = models.CharField(max_length=3, choices=CURRENCY_CHOICES, blank=False)
-    amount = models.FloatField(blank=False)
-    result = models.FloatField(blank=False)
+    amount = models.DecimalField(blank=False, decimal_places=12, max_digits=16)
+    result = models.DecimalField(blank=False, decimal_places=12, max_digits=16)
 
     class Meta:
         managed = False
